scripts/evaluator/requirements.txt
# Core dependencies
pydantic-ai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Environment management
python-dotenv>=1.0.0

# Additional database support
alembic>=1.12.0

# Environment management
python-dotenv>=1.0.0

# Validation and data processing
jsonschema>=4.17.0
python-dateutil>=2.8.2

# CLI and utilities
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking (optional)
mypy>=1.5.0 